refactor(scheme): merge alertset.go to alert.go
