+++ internal/scheme/alert.go
@@ -2,11 +2,18 @@ package scheme
 
 import (
 	"context"
+	"errors"
 	"net/url"
 	"strconv"
+	"sync"
 	"time"
 
+	"github.com/macrat/ayd/internal/ayderr"
 	api "github.com/macrat/ayd/lib-ayd"
+)
+
+var (
+	ErrInvalidAlertURL = errors.New("invalid alert URL")
 )
 
 type Alerter interface {
@@ -95,4 +102,39 @@ func (r AlertReporter) Report(_ *url.URL, rec api.Record) {
 
 func (r AlertReporter) DeactivateTarget(source *url.URL, targets ...*url.URL) {
 	r.Upstream.DeactivateTarget(source, targets...)
+}
+
+// AlertSet is a set of alerts.
+// It also implements Alerter alertinterface.
+type AlertSet []Alerter
+
+func NewAlertSet(targets []string) (AlertSet, error) {
+	alerts := make(AlertSet, len(targets))
+	errs := &ayderr.ListBuilder{What: ErrInvalidAlertURL}
+
+	for i, t := range targets {
+		var err error
+		alerts[i], err = NewAlerter(t)
+		if err != nil {
+			errs.Pushf("%s: %w", t, err)
+		}
+	}
+
+	return alerts, errs.Build()
+}
+
+// Alert of AlertSet calls all Alert methods of children parallelly.
+// This method blocks until all alerts done.
+func (as AlertSet) Alert(ctx context.Context, r Reporter, lastRecord api.Record) {
+	wg := &sync.WaitGroup{}
+
+	for _, a := range as {
+		wg.Add(1)
+		go func(a Alerter) {
+			a.Alert(ctx, r, lastRecord)
+			wg.Done()
+		}(a)
+	}
+
+	wg.Wait()
 }
